pyproject.toml
[tool.poetry]
name = "modbot"
version = "0.1.0"
description = ""
license = "AGPL"
readme = "README.md"
authors = ["Luna D Dragon <advaith.madhukar@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
rich = "^11.2.0"
python-dotenv = "^0.19.2"
hikari-lightbulb = "^2.2.0"
hikari = {extras = ["speedups"], version = "^2.0.0.dev116"}
uvloop = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
mypy = {version = "^0.941", allow-prereleases = true}
flake8 = "^4.0.1"
python-lsp-server = "^1.3.3"
pre-commit = "^2.17.0"

[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
